Some Fixes

+++ include/palette.h
@@ -97,20 +97,20 @@ void ResetPaletteFadeControl(void);
 void InvertPlttBuffer(u32 selectedPalettes);
 void TintPlttBuffer(u32 selectedPalettes, s8 r, s8 g, s8 b);
 void UnfadePlttBuffer(u32 selectedPalettes);
-void BeginFastPaletteFade(u8 submode);
+void BeginFastPaletteFade(u32 submode);
 void BeginHardwarePaletteFade(u32 blendCnt, u32 delay, u32 y, u32 targetY, u32 shouldResetBlendRegisters);
 void BlendPalettes(u32 selectedPalettes, u8 coeff, u16 color);
 void BlendPalettesFine(u32 palettes, u16 *src, u16 *dst, u32 coeff, u32 color);
-void BlendPalettesUnfaded(u32 selectedPalettes, u8 coeff, u16 color);
+void BlendPalettesUnfaded(u32 selectedPalettes, u8 coeff, u32 color);
 void BlendPalettesGradually(u32 selectedPalettes, s8 delay, u8 coeff, u8 coeffTarget, u16 color, u8 priority, u8 id);
 void TimeBlendPalette(u16 palOffset, u32 coeff, u32 blendColor);
 void TintPalette_RGB_Copy(u16 palOffset, u32 blendColor);
 void TimeMixPalettes(u32, u16 *, u16 *, struct BlendSettings *, struct BlendSettings *, u16);
 void AvgPaletteWeighted(u16 *src0, u16 *src1, u16 *dst, u16 weight0);
-void TintPalette_GrayScale(u16 *palette, u16 count);
-void TintPalette_GrayScale2(u16 *palette, u16 count);
-void TintPalette_SepiaTone(u16 *palette, u16 count);
-void TintPalette_CustomTone(u16 *palette, u16 count, u16 rTone, u16 gTone, u16 bTone);
+void TintPalette_GrayScale(u16 *palette, u32 count);
+void TintPalette_GrayScale2(u16 *palette, u32 count);
+void TintPalette_SepiaTone(u16 *palette, u32 count);
+void TintPalette_CustomTone(u16 *palette, u32 count, u16 rTone, u16 gTone, u16 bTone);
 
 static inline void SetBackdropFromColor(u32 color)
 {
